feat(place): 관광지 사진 조회 기능
- Image 그래프 탐색을 먼저 ImageMap을 모두 만들어 Batch 쿼리로 최적화
- Paging을 순수 JPA로 마이그레이션

+++ src/main/java/com/yeohangttukttak/api/global/common/PageResult.java
@@ -18,9 +18,9 @@ public class PageResult<T> {
     @AllArgsConstructor
     public static class PageInfo {
 
-        private Long page;
+        private int page;
 
-        private Long pageSize;
+        private int pageSize;
 
         private boolean hasNextPage;
